+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - 0.1.3
     - Optimized Cosine distance kernel
     - Optimized (NaN) Safe Euclidean distance kernel
+    - Fixed markedness calculation in Multiclass Breakdown
 
 - 0.1.2
     - Fixed Grid Search best hyper-parameters method